Move transform helpers to new helpers/transform.ts
Relocate translatePoint, translateProfile, transformProfile,
transformProfileAffine, rotatePointAround, mirrorProfile,
mirrorAngleAcrossLine, arcToBezierSegments, translateClamp, and
translateTab from projectStore.ts into a dedicated home module.
Slices now import these directly instead of receiving them via DI.
